% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calc_qi}
\alias{calc_qi}
\title{Calculate a quantity of interest from objects of class "post."}
\usage{
calc_qi(post, X_pred_list, prob = 0.9, qi_name = "pr", ci_type = "hpd")
}
\arguments{
\item{post}{An object of class "post" created by one of the sim_post_*()
functions.}

\item{X_pred_list}{A named list, where the names correspond to variable names
and the values correspond to values at which to set the variables. Only one
element of the list should have more than one value. If \code{qi_name =
"rr"} or \code{qi_name = "rr"} (see below), then one variable should have
exactly two values and the rest should have exactly one.}

\item{prob}{A numeric scalar in the interval (0,1) giving the target
probability content of the intervals.}

\item{qi_name}{The name of the quantity of interest to calculate. Either
\code{"pr"} (predicted probability), \code{"rr"} (risk-ratio), or
\code{"fd"} (first-difference). Defaults to \code{"pr"}.}

\item{ci_type}{The type of confidence interval to compute. Either
  \code{"hpd"} (highest posterior density) or \code{"et"} (equal-tailed).
  Defaults to \code{"hpd"}.}
}
\description{
\code{calc_qi()} calculates a quantity of interest from objects
of class "post."
}
\details{
Researchers can use this function to convert posterior simulations
of the logistic regression coefficients to quantities of interest, such as
predicted probabilities, risk-ratios, and first-differences.
}
\references{
Rainey, Carlisle. "Dealing with Separation in Logistic Regression
  Model." Working paper. Available at
  \url{http://crain.co/papers/separation.pdf}.
}

