% Generated by roxygen2 (4.0.2): do not edit by hand
\name{combine_post}
\alias{combine_post}
\title{Combine objects of class post.}
\usage{
combine_post(...)
}
\arguments{
\item{...}{Objects of class post to be combined.}
}
\description{
Combine ojects of class post.
}
\details{
Combine objects of class post.
}

