% Generated by roxygen2 (4.0.2): do not edit by hand
\name{combine_pppd}
\alias{combine_pppd}
\title{Combine objects of class pppd.}
\usage{
combine_pppd(...)
}
\arguments{
\item{...}{Objects of class pppd to be combined.}
}
\description{
Combine ojects of class pppd.
}
\details{
Combine objects of class pppd.
}

