% Generated by roxygen2 (4.0.2): do not edit by hand
\name{combine_qi}
\alias{combine_qi}
\title{Combine objects of class qi.}
\usage{
combine_qi(...)
}
\arguments{
\item{...}{Objects of class qi to be combined.}
}
\description{
Combine ojects of class qi.
}
\details{
Combine objects of class qi.
}

