% Generated by roxygen2 (4.0.2): do not edit by hand
\name{compact_hist}
\alias{compact_hist}
\title{Plots a histogram using the compact axis notation from the
 \code{compactr} package.}
\usage{
compact_hist(x, xlab = NULL, ylab = NULL, n_breaks = NULL, ht = NULL,
  xlim = NULL, counts = TRUE, bar_color = "grey70", log_scale = FALSE,
  plot = TRUE, ...)
}
\arguments{
\item{x}{A vector of values for which the histogram is desired.}

\item{xlab}{A label for the x axis. Defaults to no label.}

\item{ylab}{A label for the y axis. Defaults to "Counts" if \code{counts =
TRUE} and "Density" if \code{counts = FALSE}.}

\item{n_breaks}{one of:
\enumerate{ \item A single number giving the number of
 cells for the histogram.
 \item A character string naming an algorithm to
 compute the number of cells. Defaults to "Sturges." See documentation for
 \code{hist()} for details and other algorithms.
 \item A vector giving the
 breakpoints between histogram cells.
 \item A function to compute the vector
 of breakpoints.
 \item A function to compute the number of cells.
 }}

\item{ht}{The height of the y axis.}

\item{xlim}{The limits of the x axis.}

\item{counts}{Logical; should the heights be counts. If \code{FALSE}, then
probabilities are used instead.}

\item{bar_color}{The color of the bars. Defaults to \code{"grey70"}.}

\item{log_scale}{Logical; should the x axis be plotted on the log scale?
  Defaults to \code{FALSE}.}

\item{plot}{Logical. Should the histgram be plotted? If FALSE, the histogram is not
plotted and the data for the histogram are returned.}

\item{...}{Arguments passed to \code{eplot()} in the package \code{compactr}.}
}
\description{
\code{compact_hist()} plots a histogram using the compact axis
 notation from the \code{compactr} package.
}
\details{
Plots a histogram using the compact axis notation from the \code{compactr}
package.
}

