% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.post}
\alias{plot.post}
\title{Plot a quantity of interest}
\usage{
\method{plot}{post}(post, var_name, ci_type = "hpd", prob = 0.9,
  main = NULL, xlim = NULL, ylim = NULL, ...)
}
\arguments{
\item{post}{An object of class "post" created by one of the \code{sim_post_*()} functions.}

\item{var_name}{The name of the variable to plot.}

\item{ci_type}{The type of confidence interval to compute. Either
\code{"hpd"} (highest posterior density) or \code{"et"} (equal-tailed).
Defaults to \code{"hpd"}.}

\item{prob}{A numeric scalar in the interval (0,1) giving the target
probability content of the intervals.}

\item{main}{The title of the plot.}

\item{xlim}{The x limits (x1, x2) of the plot.}

\item{ylim}{The y limits (y1, y2) of the plot.}

\item{...}{Arguments passed to \code{eplot()} in the package \code{compactr}.}
}
\description{
\code{plot.post()} produces a density plot for a single
  coefficient from an object of class "post."
}

