% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.posts}
\alias{plot.posts}
\title{Plots densities of model coefficients using the compact axis notation from the
 \code{compactr} package.}
\usage{
\method{plot}{posts}(posts, var_name, ci_type = "hpd", prob = 0.9,
  plot_matrix_layout = NULL, xlim = NULL, ylim = NULL, ...)
}
\arguments{
\item{posts}{An object of class \code{pppds}, created by the function \link{combine_post()}.}

\item{var_name}{The name variable for the coefficient to plot.}

\item{ci_type}{The type of confidence interval to compute. Either
\code{"hpd"} (highest posterior density) or \code{"et"} (equal-tailed).
Defaults to \code{"hpd"}.}

\item{prob}{A numeric scalar in the interval (0,1) giving the target
 probability content of the intervals.}

\item{plot_matrix_layout}{A vector of the form \code{c(n_row, n_col)} to indicate the
number of rows and the number of columns in the plot matrix. Defaults to a single row.}

\item{xlim}{The x limits (x1, x2) of the plot.}

\item{ylim}{The y limits (y1, y2) of the plot.}

\item{...}{Arguments passed to \code{eplot()} in the package \code{compactr}.}
}
\description{
\code{plot.posts()} plots densities of model coefficients using the compact axis notation from the
 \code{compactr} package.
}
\details{
Plots densities of model coefficients
}

