% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.pppds}
\alias{plot.pppds}
\title{Plots histograms using the compact axis notation from the
 \code{compactr} package.}
\usage{
\method{plot}{pppds}(pppds, qi_name = "pr", n_breaks = 50,
  log_scale = FALSE, upper = NULL, lower = NULL,
  plot_matrix_layout = NULL, arrow_ht = 0.2, ...)
}
\arguments{
\item{pppds}{An object of class \code{pppds}, created by the function \link{combine_pppd()}.}

\item{qi_name}{The name of the quantity of interest to plot. Defaults to \code{"pr"}.
The other options are \code{"fd"} and \code{"rr"}.
\enumerate{ \item A single number giving the number of
 cells for the histogram. Defaults to 50. For examining the partial prior
 predictive distritution, too many breaks are preferable to too few.
 \item A character string naming an algorithm to
 compute the number of cells. See documentation for
 \code{hist()} for details and other algorithms.
 \item A vector giving the
 breakpoints between histogram cells.
 \item A function to compute the vector
 of breakpoints.
 \item A function to compute the number of cells.
 }}

\item{log_scale}{Logical; should the quantity of interest be plotted on the log scale?
  Defaults to \code{FALSE}. I recommend setting to \code{TRUE} if the distribution
  has a long right tail, as can happen with risk-ratios in certain situations.}

\item{upper}{The upper bound of quantity of interest to plot.}

\item{lower}{The lower bound of quantity of interest to plot.}

\item{plot_matrix_layout}{A vector of the form \code{c(n_row, n_col)} to indicate the
number of rows and the number of columns in the plot matrix. Defaults to a single row.}

\item{arrow_ht}{The height of the information about truncation. Defaults to
0.2. The bottom of the plot is 0.0 and the top is 1.0.}

\item{...}{Arguments passed to \code{compact_hist()}.}
}
\description{
\code{plot.pppds()} plots histograms using the compact axis
 notation from the \code{compactr} package.
}
\details{
Plots a histograms using the compact axis notation from the \code{compactr}
package.
}

