% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.qi}
\alias{plot.qi}
\title{Plot a quantity of interest}
\usage{
\method{plot}{qi}(qi, x = NULL, xlim = NULL, ylim = NULL, ...)
}
\arguments{
\item{qi}{An object of class "qi" created by the \code{calc_qi()} function.}

\item{x}{A numeric variable with which the qi changes. Usually, this is the
element of \code{X_pred_list} given to \code{calc_qi()} that varies.}

\item{xlim}{The x limits (x1, x2) of the plot.}

\item{ylim}{The y limits (y1, y2) of the plot.}

\item{...}{Arguments passed to \code{eplot()} in the package \code{compactr}.}
}
\description{
\code{plot.qi()} plots an object of class "qi," eithering giving a
density plot of the quantity of interest is for a single qi (i.e., a single
predicted probability, risk-ratio, or first-difference) or a plot of the
predicted probability and credible intervals if the qi is a range of predicted
probabilities.
}

