% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.qis}
\alias{plot.qis}
\title{Plots quantities of interest using the compact axis notation from the
 \code{compactr} package.}
\usage{
\method{plot}{qis}(qis, x_vals = NULL, plot_matrix_layout = NULL,
  xlim = NULL, ylim = NULL, ...)
}
\arguments{
\item{qis}{An object of class \code{pppds}, created by the function \link{combine_qi()}.}

\item{x_vals}{The x values across which the quantities of interest vary}

\item{plot_matrix_layout}{A vector of the form \code{c(n_row, n_col)} to indicate the
number of rows and the number of columns in the plot matrix. Defaults to a single row.}

\item{xlim}{The x limits (x1, x2) of the plot.}

\item{ylim}{The y limits (y1, y2) of the plot.}

\item{...}{Arguments passed to \code{eplot()} in the package \code{compactr}.}
}
\description{
\code{plot.qis()} plots quantities of interest using the compact axis notation from the
 \code{compactr} package.
}
\details{
Plots quantities of interest
}

