% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_hyp_test}
\alias{plot_hyp_test}
\title{Plots a histogram using the compact axis notation from the
 \code{compactr} package.}
\usage{
plot_hyp_test(posts, var_name, research_hyp, xlab = NULL, ...)
}
\arguments{
\item{posts}{An object of class \code{pppds}, created by the function
\link{combine_post()}.}

\item{var_name}{The name of the variable of interest.}

\item{research_hyp}{The research hypothesis. Either that the coefficient for the variable of interest is
positive (\code{"positive"}, \code{"p"}, or \code{"+"}) or negative (\code{"negative"}, \code{"n"}, or \code{"-"})}

\item{xlab}{The label for the x axis.}

\item{...}{Arguments passed to \code{compact_hist()}.}
}
\description{
\code{plot_hyp_tests()} plots the probability of the research
 hypothesis across several models the compact axis notation from the
 \code{compactr} package.
}
\details{
Plots a histogram using the compact axis notation from the \code{compactr}
package.
}

