% Generated by roxygen2 (4.0.2): do not edit by hand
\name{post_to_bugs}
\alias{post_to_bugs}
\title{Convert an object of class "post" to an object of class "bugs"}
\usage{
post_to_bugs(post)
}
\arguments{
\item{post}{An object of class "post" created by one of the post_sim_*() functions.}
}
\description{
\code{post_to_bugs()} converts an object of class "post" to an
object of class "bugs," which makes a few methods for monitoring convergence available.
}
\details{
Convert an object of class "post" to an object of class "bugs"
}

