% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.post}
\alias{print.post}
\title{Print summary of the partial prior predictive distribution.}
\usage{
print.post(post)
}
\arguments{
\item{post}{An object created by the functions \code{sim_post_*()} with class \code{post} or \code{post_gelman}.}
}
\description{
\code{print.post()} and \code{print.post_gelman()} print a summary for object of the class \code{post} and \code{post_gelman} created by the functions \code{sim_post_*()}.
}
\details{
Print summary of the partial prior predictive distribution.
}

