% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.pppd}
\alias{print.pppd}
\title{Print summary of the partial prior predictive distribution.}
\usage{
\method{print}{pppd}(pppd, probs = c(0.01, 0.05, 0.25, 0.5, 0.75, 0.95, 0.99),
  round_to = 3)
}
\arguments{
\item{pppd}{An object of class pppd.}

\item{probs}{A numeric vector of probabilities with values in [0,1] a which to compute the quantile.}

\item{round_to}{The number of decimal place to round the results.}
}
\description{
\code{print.pppd()} prints a summary for object of the class pppd
}
\details{
Print summary of the partial prior predictive distribution.
}

