% Generated by roxygen2 (4.0.2): do not edit by hand
\name{set_rest_at_median}
\alias{set_rest_at_median}
\title{Vary one variable and set the rest at their medians}
\usage{
set_rest_at_median(formula, data, var_name, var_values)
}
\arguments{
\item{formula}{A model formula.}

\item{data}{A data frame.}

\item{var_name}{The name of the variable to be varied.}

\item{var_values}{The values that \code{var_name} should take.}
}
\description{
\code{set_rest_at_median} returns a list of the median of every
 variables with onve variable taking on a given range of values. This
 function is useful for generating matrices for computing quantities of
 interst, such as predicted probabilities, risk-ratios, and
 first-differences.
}
\details{
Vary one variable and set the rest at their medians
}

