% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sim_post_jeffreys}
\alias{sim_post_jeffreys}
\title{Obtain posterior simulations using Jeffreys' prior}
\usage{
sim_post_jeffreys(formula, data, n_sims = 1000, n_burnin = n_sims/2,
  n_chains = 3, n_thin = 1, n_cores = n_chains, tune = 1)
}
\arguments{
\item{formula}{A logistic regression model.}

\item{data}{A data frame.}

\item{n_sims}{The number of simulations after the burn-in period.}

\item{n_burnin}{The number of burn-in iterations for the sample.}

\item{n_chains}{The number of MCMC chains being run.}

\item{n_thin}{The thinning interval used in the simulation. The number of MCMC
iterations must be divisible by this value.}

\item{n_cores}{The number of MCMC cores. Defaults to the number of chains.}

\item{tune}{The tuning parameter for the Metropolis sampling. Can be either a
positive scalar or a (k+1)-vector, where k is the number of variables in the
model. Presently passed to \code{MCMCmetrop1R}.}
}
\description{
\code{sim_post_jeffreys()} uses a Metropolis algorithm to obtain posterior
simulations using Jeffreys' prior.
}
\references{
Firth, David. 1993. "Bias Reduction of Maximum Likelihood
  Estimates." Biometrika 80(1):27–38.

Heinze, Georg, and Michael Schemper. 2002. "A Solution to the
  Problem of Separation in Logistic Regression." Statistics in Medicine
  21(16):2409–2419.

Jeffreys, H. 1946. "An Invariant Form of the Prior Probability in
  Estimation Problems." Proceedings of the Royal Society of London, Series A
  186(1007):453–461.

Poirier, Dale. 1994. "Jeffreys’ Prior for Logit Models." Journal
  of Econometrics 63(2):327–339.

Zorn, Christopher. 2005. "A Solution to Separation in Binary
  Response Models." Political Analysis 13(2):157–170.
}

