% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sim_post_normal}
\alias{sim_post_normal}
\title{Obtain posterior simulations using a normal prior on the separating variable.}
\usage{
sim_post_normal(formula, data, sep_var, sd = 4.5, n_sims = 1000,
  n_burnin = n_sims/2, n_thin = 1, n_chains = 3, n_cores = n_chains,
  tune = 1)
}
\arguments{
\item{formula}{A logistic regression model.}

\item{data}{A data frame.}

\item{sep_var}{The name of the separating variable.}

\item{sd}{The standard deviation of the normal prior on the coefficient for
the separating variable.}

\item{n_sims}{The number of simulations after the burn-in period.}

\item{n_burnin}{The number of burn-in iterations for the sample.}

\item{n_thin}{The thinning interval used in the simulation. The number of MCMC
iterations must be divisible by this value.}

\item{n_chains}{The number of MCMC chains being run.}

\item{n_cores}{The number of MCMC cores. Defaults to the number of chains.}

\item{tune}{The tuning parameter for the Metropolis sampling. Can be either a
positive scalar or a (k+1)-vector, where k is the number of variables in the
model. Presently passed to \code{MCMCmetrop1R}.}
}
\description{
\code{sim_post_normal()} uses a Metropolis algorithm to obtain posterior
simulations using a normal prior on the separating variable and improper, uinform
priors on the remaining variables.
}

