% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.post}
\alias{print.post}
\title{Print summary of the partial prior predictive distribution.}
\usage{
\method{print}{post}(post, digits = 2, prob = 0.9)
}
\arguments{
\item{post}{An object created by the functions \code{sim_post_*()} with class \code{post}.}

\item{digits}{The number of digits for printing.}

\item{prob}{A numeric scalar in the interval (0,1) giving the target probability content of the intervals.}
}
\description{
\code{print.post()} prints a summary for objects of the class \code{post} created by the functions \code{sim_post_*()}.
}
\details{
Print summary of the partial prior predictive distribution.
}

