% Generated by roxygen2 (4.0.2): do not edit by hand
\name{set_at_median}
\alias{set_at_median}
\title{Set variables at their medians}
\usage{
set_at_median(x)
}
\arguments{
\item{x}{Either a data frame or a matrix with column names.}
}
\description{
\code{set_at_median} returns a list of the median of every
 variables. This function is useful for generating matrices for computing
 quantities of interst, such as predicted probabilities, risk-ratios, and
 first-differences.
}
\details{
Set variables at their medians
}

